+++ src/main/java/Character/Person.java
@@ -1,7 +1,6 @@
-package practice250610;
+package Character;
 
 import java.time.LocalDate;
-import java.util.Date;
 
 public class Person {
     private String name;
