+++ src/test/java/Character/ClericTest.java
@@ -1,6 +1,5 @@
-package practice250610;
+package Character;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
